apps/rag/requirements.txt
langchain
langchain-community
llama-index
sentence-transformers
chromadb
faiss-cpu
transformers
accelerate
bitsandbytes
evaluate
scikit-learn
peft
tqdm
pandas
datasets
matplotlib
streamlit
PyMuPDF
unstructured
pytest
python-dotenv
streamlit-authenticator==0.2.2